fix: fix the gallery loader display on initial page load
